+++ src/Helpers/ProjectHelpers.cs
@@ -1,0 +1,42 @@
+using Microsoft.VisualStudio;
+using Microsoft.VisualStudio.ComponentModelHost;
+using Microsoft.VisualStudio.Editor;
+using Microsoft.VisualStudio.Shell;
+using Microsoft.VisualStudio.Text;
+using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.TextManager.Interop;
+
+namespace IgnoreFiles
+{
+    class ProjectHelpers
+    {
+
+        public static ITextBuffer GetCurentTextBuffer()
+        {
+            return GetCurentTextView().TextBuffer;
+        }
+
+        public static IWpfTextView GetCurentTextView()
+        {
+            var componentModel = GetComponentModel();
+            if (componentModel == null) return null;
+            var editorAdapter = componentModel.GetService<IVsEditorAdaptersFactoryService>();
+
+            return editorAdapter.GetWpfTextView(GetCurrentNativeTextView());
+        }
+
+        private static IVsTextView GetCurrentNativeTextView()
+        {
+            var textManager = (IVsTextManager)ServiceProvider.GlobalProvider.GetService(typeof(SVsTextManager));
+
+            IVsTextView activeView = null;
+            ErrorHandler.ThrowOnFailure(textManager.GetActiveView(1, null, out activeView));
+            return activeView;
+        }
+
+        private static IComponentModel GetComponentModel()
+        {
+            return (IComponentModel)Package.GetGlobalService(typeof(SComponentModel));
+        }
+    }
+}