+++ CHANGELOG.md
@@ -6,9 +6,10 @@
 - [x] Mark paths that doesn't hit any files
 - [x] Hover QuickInfo for file match details
 - [x] Support globbing/minimatch
-- [ ] Command to remove all lines that matches nothing
+- [x] Command to remove all lines that matches nothing
 - [ ] Intellisense for file paths
 - [ ] Options page
+- [ ] Package load on .ignore file open
 
 Features that have a checkmark are complete and available for
 download in the